# feat(stdlib/universe): implement schema mutation functions without performing any copies