--- a/stdlib/universe/schema_functions.go
+++ b/stdlib/universe/schema_functions.go
@@ -3,11 +3,15 @@ package universe
 import (
 	"context"
 
+	"github.com/apache/arrow/go/arrow/array"
 	"github.com/influxdata/flux"
+	"github.com/influxdata/flux/arrow"
 	"github.com/influxdata/flux/codes"
 	"github.com/influxdata/flux/execute"
 	"github.com/influxdata/flux/internal/errors"
+	"github.com/influxdata/flux/internal/execute/table"
 	"github.com/influxdata/flux/interpreter"
+	"github.com/influxdata/flux/memory"
 	"github.com/influxdata/flux/plan"
 	"github.com/influxdata/flux/runtime"
 	"github.com/influxdata/flux/semantic"
@@ -103,8 +107,8 @@ func init() {
 		r.Register()
 	}
 
-	plan.RegisterProcedureSpec(SchemaMutationKind, newSchemaMutationProcedure, SchemaMutationOps...)
-	execute.RegisterTransformation(SchemaMutationKind, createSchemaMutationTransformation)
+	plan.RegisterProcedureSpec(SchemaMutationKind, newDualImplSpec(newSchemaMutationProcedure), SchemaMutationOps...)
+	execute.RegisterTransformation(SchemaMutationKind, createDualImplTf(createSchemaMutationTransformation, createDeprecatedSchemaMutationTransformation))
 }
 
 func createRenameOpSpec(args flux.Arguments, a *flux.Administration) (flux.OperationSpec, error) {
@@ -393,7 +397,7 @@ func (s *SchemaMutationProcedureSpec) Kind() plan.ProcedureKind {
 
 func (s *SchemaMutationProcedureSpec) Copy() plan.ProcedureSpec {
 	newMutations := make([]SchemaMutation, len(s.Mutations))
-	for i, m := range newMutations {
+	for i, m := range s.Mutations {
 		newMutations[i] = m.Copy()
 	}
 
@@ -415,43 +419,40 @@ func newSchemaMutationProcedure(qs flux.OperationSpec, pa plan.Administration) (
 
 type schemaMutationTransformation struct {
 	d        execute.Dataset
-	cache    execute.TableBuilderCache
+	cache    table.BuilderCache
 	ctx      context.Context
 	mutators []SchemaMutator
 }
 
 func createSchemaMutationTransformation(id execute.DatasetID, mode execute.AccumulationMode, spec plan.ProcedureSpec, a execute.Administration) (execute.Transformation, execute.Dataset, error) {
-	cache := execute.NewTableBuilderCache(a.Allocator())
-	d := execute.NewDataset(id, mode, cache)
-
-	t, err := NewSchemaMutationTransformation(a.Context(), spec, d, cache)
-	if err != nil {
-		return nil, nil, err
-	}
-	return t, d, nil
-}
-
-func NewSchemaMutationTransformation(ctx context.Context, spec plan.ProcedureSpec, d execute.Dataset, cache execute.TableBuilderCache) (*schemaMutationTransformation, error) {
 	s, ok := spec.(*SchemaMutationProcedureSpec)
 	if !ok {
-		return nil, errors.Newf(codes.Internal, "invalid spec type %T", spec)
+		return nil, nil, errors.Newf(codes.Internal, "invalid spec type %T", spec)
 	}
+	return NewSchemaMutationTransformation(a.Context(), s, id, a.Allocator())
+}
 
-	mutators := make([]SchemaMutator, len(s.Mutations))
-	for i, mutation := range s.Mutations {
+func NewSchemaMutationTransformation(ctx context.Context, spec *SchemaMutationProcedureSpec, id execute.DatasetID, mem *memory.Allocator) (execute.Transformation, execute.Dataset, error) {
+	mutators := make([]SchemaMutator, len(spec.Mutations))
+	for i, mutation := range spec.Mutations {
 		m, err := mutation.Mutator()
 		if err != nil {
-			return nil, err
+			return nil, nil, err
 		}
 		mutators[i] = m
 	}
 
-	return &schemaMutationTransformation{
-		d:        d,
-		cache:    cache,
+	t := &schemaMutationTransformation{
+		cache: table.BuilderCache{
+			New: func(key flux.GroupKey) table.Builder {
+				return table.NewBufferedBuilder(key, mem)
+			},
+		},
 		mutators: mutators,
 		ctx:      ctx,
-	}, nil
+	}
+	t.d = table.NewDataset(id, &t.cache)
+	return t, t.d, nil
 }
 
 func (t *schemaMutationTransformation) Process(id execute.DatasetID, tbl flux.Table) error {
@@ -463,38 +464,30 @@ func (t *schemaMutationTransformation) Process(id execute.DatasetID, tbl flux.Ta
 		}
 	}
 
-	builder, created := t.cache.TableBuilder(ctx.Key())
-	if created {
-		for _, c := range ctx.Cols() {
-			_, err := builder.AddCol(c)
-			if err != nil {
-				return err
-			}
-		}
-	} else {
-		// We are appending to an existing table, due to dropping columns in the group key.
-		// Make sure that tables are compatible.
-		if len(ctx.Cols()) != len(builder.Cols()) {
-			key := builder.Key().String()
-			return errors.New(codes.Invalid, "requested operation merges tables with different numbers of columns for group key "+key)
-		}
-		for i, cm := range ctx.Cols() {
-			bcm := builder.Cols()[i]
-			if cm != bcm {
-				key := builder.Key().String()
-				return errors.New(codes.Invalid, "requested operation merges tables with different schemas for group key "+key)
+	mutTable, err := t.mutateTable(tbl, ctx)
+	if err != nil {
+		tbl.Done()
+		return err
+	}
+	builder, _ := table.GetBufferedBuilder(mutTable.Key(), &t.cache)
+	return mutTable.Do(builder.AppendBuffer)
+}
+
+func (t *schemaMutationTransformation) mutateTable(in flux.Table, ctx *BuilderContext) (flux.Table, error) {
+	// Check the schema for columns with the same name.
+	cols := ctx.Cols()
+	for i, c := range cols {
+		for j := range cols[:i] {
+			if cols[j].Label == c.Label {
+				return nil, errors.Newf(codes.FailedPrecondition, "column %d and %d have the same name (%q) which is not allowed", j, i, c.Label)
 			}
 		}
 	}
 
-	return tbl.Do(func(cr flux.ColReader) error {
-		for i := 0; i < cr.Len(); i++ {
-			if err := execute.AppendMappedRecordWithNulls(i, cr, builder, ctx.ColMap()); err != nil {
-				return err
-			}
-		}
-		return nil
-	})
+	return &mutateTable{
+		Table: in,
+		ctx:   ctx,
+	}, nil
 }
 
 func (t *schemaMutationTransformation) RetractTable(id execute.DatasetID, key flux.GroupKey) error {
@@ -511,4 +504,28 @@ func (t *schemaMutationTransformation) UpdateProcessingTime(id execute.DatasetID
 
 func (t *schemaMutationTransformation) Finish(id execute.DatasetID, err error) {
 	t.d.Finish(err)
+}
+
+type mutateTable struct {
+	flux.Table
+	ctx *BuilderContext
+}
+
+func (m *mutateTable) Key() flux.GroupKey   { return m.ctx.Key() }
+func (m *mutateTable) Cols() []flux.ColMeta { return m.ctx.Cols() }
+
+func (m *mutateTable) Do(f func(flux.ColReader) error) error {
+	return m.Table.Do(func(cr flux.ColReader) error {
+		indices := m.ctx.ColMap()
+		buffer := &arrow.TableBuffer{
+			GroupKey: m.Key(),
+			Columns:  m.Cols(),
+			Values:   make([]array.Interface, len(indices)),
+		}
+		for j, idx := range indices {
+			buffer.Values[j] = table.Values(cr, idx)
+			buffer.Values[j].Retain()
+		}
+		return f(buffer)
+	})
 }