This changes the schema mutation functions to mutate the table in place
without requiring any copies to the data. The schema for the table will
be mutated by wrapping the table and manipulating the schema on the
`flux.Table` object instead of requiring a copy of the data. The retain
functions for arrays are used to facilitate this.

Then it uses the buffered table builder to append the buffers to each
other to perform the regrouping which mimics the same code that group
uses.

Fixes #2879.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written