@jsternberg 2020-06-26
Benchmark comparison:

```
jsternberg:[~/g/s/g/i/flux] $ (feat/schema-mutator-nocopy $) benchcmp old.txt new.txt
benchmark                             old ns/op       new ns/op     delta
BenchmarkKeep_Values/1000-12          14865928226     116163523     -99.22%
BenchmarkDrop_Values/1000-12          23296183178     162736991     -99.30%
BenchmarkRename_Values/1000-12        45063227845     254866548     -99.43%
BenchmarkDuplicate_Values/1000-12     52520034736     229051272     -99.56%

benchmark                             old allocs     new allocs     delta
BenchmarkKeep_Values/1000-12          120431153      661457         -99.45%
BenchmarkDrop_Values/1000-12          238300834      641600         -99.73%
BenchmarkRename_Values/1000-12        362812044      698161         -99.81%
BenchmarkDuplicate_Values/1000-12     420927359      573047         -99.86%

benchmark                             old bytes       new bytes     delta
BenchmarkKeep_Values/1000-12          10144526088     43812368      -99.57%
BenchmarkDrop_Values/1000-12          14836211840     40638193      -99.73%
BenchmarkRename_Values/1000-12        18452306344     43796336      -99.76%
BenchmarkDuplicate_Values/1000-12     21359743736     43831057      -99.79%
```