@jsternberg 2020-06-29
I readded the tests for the deprecated code that ran the transformation. I'm a little hesitant to do more like duplicate the end to end tests because they're going to have to be removed and I think the work to create and maintain them will be greater than the deprecated implementation will exist. That code isn't going to change so if it was previously working it will likely continue working.